GroupSiwen/requirements.txt
google-cloud-bigquery
pendulum
